+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Added
 - Display the JWT expiration Unix timestamp in `sensuctl config view`.
 - Added the 'sensu-backend init' subcommand.
+- Added a new flag, --etcd-client-urls, which should be used with sensu-backend
+when it is not operating as an etcd member. The flag is also used by the new
+sensu-backend init tool.
 
 ### Fixed
 - Add a timeout to etcd requests when retrieving the nodes health.